backend/house_price_predictor/requirements.txt
django
djangorestframework
pandas
numpy
scikit-learn
django-cors-headers